teste com regra de tempo 3

+++ src/main/java/com/example/controller/response/PautaDto.java
@@ -16,5 +16,5 @@ public class PautaDto {
 
     private String nome;
 
-    private Map<String, Integer> resultado;
+    private Map<String, Long> resultado;
 }
